requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
dj-database-url==3.1.2
Django==6.0.3
django-crispy-forms==2.6
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
whitenoise==6.12.0
